# #87 Iterator common item implements