--- a/bench/filter-compare-1.txt
+++ b/bench/filter-compare-1.txt
@@ -2,9 +2,9 @@ goos: linux
 goarch: amd64
 pkg: github.com/meian/gcf/bench
 cpu: Intel(R) Core(TM) i5-8250U CPU @ 1.60GHz
-BenchmarkFilter_Compare/filter         	 2338778	      2547 ns/op	     144 B/op	       4 allocs/op
-BenchmarkFilter_Compare/if-func        	 9481165	       620.8 ns/op	       0 B/op	       0 allocs/op
-BenchmarkFilter_Compare/if-inline      	155062731	        38.53 ns/op	       0 B/op	       0 allocs/op
-BenchmarkFilter_Compare/chan           	   76498	     78090 ns/op	     384 B/op	       6 allocs/op
+BenchmarkFilter_Compare/filter         	 2237217	      2683 ns/op	     192 B/op	       4 allocs/op
+BenchmarkFilter_Compare/if-func        	 9417493	       624.2 ns/op	       0 B/op	       0 allocs/op
+BenchmarkFilter_Compare/if-inline      	155073327	        38.57 ns/op	       0 B/op	       0 allocs/op
+BenchmarkFilter_Compare/chan           	   76040	     79738 ns/op	     384 B/op	       6 allocs/op
 PASS
-ok  	github.com/meian/gcf/bench	31.790s
+ok  	github.com/meian/gcf/bench	32.027s