Closes #87 

### What changed

- Add Iterator common items.
    - with current and done
- Apply iteratorItem to each Iterator
    - First step in each Iterator is unified to check done.
    - Prevented unnecessary setting of variables in the initialization of each Iterator.

The change is refactoring, so test is no changed except leaked test added.

### Others

----

### Checklist

- [x] **Issue is linked.**
    - It is not needed if simple edit such as corrections of words, etc.
- [ ] **Test is added.**
    - If fixed bug, add case on existing test or add new test.
    - If add new feature, add new test.
    - Add benchmark test, as needed.
- [x] **Test coverage rate has no decreased.**
    - If rate is decreased, explain why in What changed.