@meian 2022-02-06
Filterのパフォーマンス比較  
特に差はなさそう

```
$ benchstat <(git show HEAD^:bench/filter-volumes.txt) bench/filter-volumes.txt 
name                                    old time/op    new time/op    delta
Filter_Volumes/slice=10/filter=1-8         202ns ± 0%     201ns ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=10/filter=5-8         703ns ± 0%     755ns ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=10/filter=10-8       1.32µs ± 0%    1.38µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=10/filter=100-8      11.5µs ± 0%    11.7µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=100/filter=1-8       1.01µs ± 0%    1.04µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=100/filter=5-8       4.65µs ± 0%    4.93µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=100/filter=10-8      8.55µs ± 0%    8.68µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=100/filter=100-8     70.8µs ± 0%    69.2µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=1000/filter=1-8      9.08µs ± 0%    9.19µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=1000/filter=5-8      45.1µs ± 0%    45.9µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=1000/filter=10-8     80.2µs ± 0%    81.0µs ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=1000/filter=100-8     684µs ± 0%     643µs ± 0%   ~     (p=1.000 n=1+1)

name                                    old alloc/op   new alloc/op   delta
Filter_Volumes/slice=10/filter=1-8         80.0B ± 0%     96.0B ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=10/filter=5-8          208B ± 0%      288B ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=10/filter=10-8         368B ± 0%      528B ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=10/filter=100-8      3.25kB ± 0%    4.85kB ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=100/filter=1-8        80.0B ± 0%     96.0B ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=100/filter=5-8         208B ± 0%      288B ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=100/filter=10-8        368B ± 0%      528B ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=100/filter=100-8     3.25kB ± 0%    4.85kB ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=1000/filter=1-8       80.0B ± 0%     96.0B ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=1000/filter=5-8        208B ± 0%      288B ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=1000/filter=10-8       368B ± 0%      528B ± 0%   ~     (p=1.000 n=1+1)
Filter_Volumes/slice=1000/filter=100-8    3.25kB ± 0%    4.85kB ± 0%   ~     (p=1.000 n=1+1)

name                                    old allocs/op  new allocs/op  delta
Filter_Volumes/slice=10/filter=1-8          2.00 ± 0%      2.00 ± 0%   ~     (all equal)
Filter_Volumes/slice=10/filter=5-8          6.00 ± 0%      6.00 ± 0%   ~     (all equal)
Filter_Volumes/slice=10/filter=10-8         11.0 ± 0%      11.0 ± 0%   ~     (all equal)
Filter_Volumes/slice=10/filter=100-8         101 ± 0%       101 ± 0%   ~     (all equal)
Filter_Volumes/slice=100/filter=1-8         2.00 ± 0%      2.00 ± 0%   ~     (all equal)
Filter_Volumes/slice=100/filter=5-8         6.00 ± 0%      6.00 ± 0%   ~     (all equal)
Filter_Volumes/slice=100/filter=10-8        11.0 ± 0%      11.0 ± 0%   ~     (all equal)
Filter_Volumes/slice=100/filter=100-8        101 ± 0%       101 ± 0%   ~     (all equal)
Filter_Volumes/slice=1000/filter=1-8        2.00 ± 0%      2.00 ± 0%   ~     (all equal)
Filter_Volumes/slice=1000/filter=5-8        6.00 ± 0%      6.00 ± 0%   ~     (all equal)
Filter_Volumes/slice=1000/filter=10-8       11.0 ± 0%      11.0 ± 0%   ~     (all equal)
Filter_Volumes/slice=1000/filter=100-8       101 ± 0%       101 ± 0%   ~     (all equal)
```