@meian 2022-02-06
Rangeのベンチマーク  
こちらも特に差がなさそう

```
$ benchstat <(git show HEAD^:bench/range.txt) bench/range.txt 
name                                               old time/op    new time/op    delta
Range/start=1/end=1000000/step=1/count=1000000-8     2.40ms ± 0%    2.42ms ± 0%   ~     (p=1.000 n=1+1)
Range/start=1/end=1000000/step=10/count=100000-8      239µs ± 0%     242µs ± 0%   ~     (p=1.000 n=1+1)
Range/start=1/end=1000000/step=100/count=10000-8     24.0µs ± 0%    24.1µs ± 0%   ~     (p=1.000 n=1+1)
Range/start=1000000/end=1/step=-1/count=1000000-8    2.40ms ± 0%    2.39ms ± 0%   ~     (p=1.000 n=1+1)
Range/start=1000000/end=1/step=-10/count=100000-8     240µs ± 0%     239µs ± 0%   ~     (p=1.000 n=1+1)
Range/start=1000000/end=1/step=-100/count=10000-8    24.0µs ± 0%    24.0µs ± 0%   ~     (p=1.000 n=1+1)

name                                               old alloc/op   new alloc/op   delta
Range/start=1/end=1000000/step=1/count=1000000-8      32.0B ± 0%     48.0B ± 0%   ~     (p=1.000 n=1+1)
Range/start=1/end=1000000/step=10/count=100000-8      32.0B ± 0%     48.0B ± 0%   ~     (p=1.000 n=1+1)
Range/start=1/end=1000000/step=100/count=10000-8      32.0B ± 0%     48.0B ± 0%   ~     (p=1.000 n=1+1)
Range/start=1000000/end=1/step=-1/count=1000000-8     32.0B ± 0%     48.0B ± 0%   ~     (p=1.000 n=1+1)
Range/start=1000000/end=1/step=-10/count=100000-8     32.0B ± 0%     48.0B ± 0%   ~     (p=1.000 n=1+1)
Range/start=1000000/end=1/step=-100/count=10000-8     32.0B ± 0%     48.0B ± 0%   ~     (p=1.000 n=1+1)

name                                               old allocs/op  new allocs/op  delta
Range/start=1/end=1000000/step=1/count=1000000-8       1.00 ± 0%      1.00 ± 0%   ~     (all equal)
Range/start=1/end=1000000/step=10/count=100000-8       1.00 ± 0%      1.00 ± 0%   ~     (all equal)
Range/start=1/end=1000000/step=100/count=10000-8       1.00 ± 0%      1.00 ± 0%   ~     (all equal)
Range/start=1000000/end=1/step=-1/count=1000000-8      1.00 ± 0%      1.00 ± 0%   ~     (all equal)
Range/start=1000000/end=1/step=-10/count=100000-8      1.00 ± 0%      1.00 ± 0%   ~     (all equal)
Range/start=1000000/end=1/step=-100/count=10000-8      1.00 ± 0%      1.00 ± 0%   ~     (all equal)
```